Outsource render debug and other stuff
